fix(lexical-editor): update editor state via UpdateStatePlugin
