+++ packages/lexical-editor/src/plugins/LexicalUpdateStatePlugin/index.ts
@@ -1,1 +1,1 @@
-export * from "./LexicalUpdateStatePlugin";
+export * from "./UpdateStatePlugin";
